feat: updated notebook updated scrape_tool.py
Updated the notebook: Using a PKL file with more data. Adding cleaning and lemmatization of the reviews. Modified baseline model to have a larger vocabulary size
and with larger vectors in the word embedding.

Updated scrape_tool.py to scrape at a slower rate. Updated the name of the movie to pick up from.