feat: added exception handling to MyMoviewReview. Doing baselining with a RF model. Built an improved CNN alternative. Experimented with LSTMs and CNNs
Added more exception handling to MyMoviewReview
Did a lot of experimenting with models.
Tried to build a bigger CNN (more convolutional layers). The best I could get was 33% accuracy with a
bigger model. Tried a lot of regularization, playing with EarlyStopping() and
ReduceLROnPlateau() but did not get better results

Added a random forest classifier as the baseline
Did a few experiments with LSTMs that did not pan out
Finally built a CNN model with improved performance (35% accuracy)

Did some error analysis of the CNN with 35% accuracy.

updates to scrape_tool.py are just to get it to pick after a few crashes